Data layer and IIndexerSource plugin updates:
1. Drop indexes before sync, then re-create after. This speeds up the
indexing process.

2. 'external_id' field now required for all tracks.

3. Added IIndexerSource::HasStableIds(). We use this as a hint to not
remove orphaned tracks from existing playlists when re-scanning.

4. Removed 'track_id' from 'playlist_tracks' table, and added
'track_external_id' and 'source_id' instead.

5. Indexer will only scan for plugins once now, during initialization.

+++ src/core/db/Connection.h
@@ -65,7 +65,7 @@ namespace musik { namespace core { namespace db {
             int Close();
             int Execute(const char* sql);
             int Execute(const wchar_t* sql);
-            int LastInsertedId();
+            unsigned long long LastInsertedId();
             int LastModifiedRowCount();
 
             void Interrupt();
